src/Demo.Core/Math.cs
namespace Demo.Core
{
  /// <summary>
  /// It contains the helper functions of the application.
  /// </summary>
  public class Math
  {

    /// <summary>
    /// Add the two numbers
    /// </summary>
    /// <param name="number1"></param>
    /// <param name="number2"></param>
    /// <returns>decimal</returns>
    public decimal Addition(decimal number1, decimal number2) => number1 + number2;


    /// <summary>
    /// Multiply the two numbers
    /// </summary>
    /// <param name="number1"></param>
    /// <param name="number2"></param>
    /// <returns>decimal</returns>
    public decimal Multiplication(decimal number1, decimal number2) => number1 * number2;
  }
}
